# projects/gin-advanced/docs/swagger.yaml
basePath: /
definitions:
  api.AccessLogsResponse:
    properties:
      list: {}
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
...
  api.CaptchaResponse:
    properties:
      captcha_id:
        type: string
      image:
        type: string
    type: object
  api.CleanAccessLogsRequest:
    properties:
      days:
        type: integer
    type: object
  api.DeleteSystemLogsRequest:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  api.SystemLogsResponse:
    properties:
      list: {}
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
...
  api.VerifyCaptchaRequest:
    properties:
      captcha_id:
        type: string
      code:
        type: string
    required:
    - captcha_id
    - code
    type: object
...
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Gin-Advanced RESTful API 文档
  title: Gin-Advanced API
  version: "1.0"
paths:
  /admin/logs/access:
    delete:
      consumes:
      - application/json
      description: 删除指定天数之前的访问日志
      parameters:
      - description: 保留天数
        in: body
        name: body
        schema:
          $ref: '#/definitions/api.CleanAccessLogsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 清理访问日志
      tags:
      - 访问日志
    get:
      description: 分页获取访问日志，支持按用户、请求方法、路径等条件筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 用户ID
        in: query
        name: user_id
        type: integer
      - description: 请求方法
        in: query
        name: method
        type: string
      - description: 请求路径
        in: query
        name: path
        type: string
      - description: IP地址
        in: query
        name: ip
        type: string
      - description: 状态码
        in: query
        name: status_code
        type: integer
      - description: 开始时间戳
        in: query
        name: start_time
        type: integer
      - description: 结束时间戳
        in: query
        name: end_time
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.AccessLogsResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取访问日志列表
      tags:
      - 访问日志
  /admin/logs/access/stats:
    get:
      description: 统计指定时间范围内的访问日志数量和趋势
      parameters:
      - description: 开始时间戳
        in: query
        name: start_time
        type: integer
      - description: 结束时间戳
        in: query
        name: end_time
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取访问统计
      tags:
      - 访问日志
  /admin/logs/system:
    delete:
      consumes:
      - application/json
      description: 批量删除指定的系统日志
      parameters:
      - description: 日志ID列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.DeleteSystemLogsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除系统日志
      tags:
      - 系统日志
    get:
      description: 分页获取系统日志，支持按用户、模块、操作等条件筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 用户ID
        in: query
        name: user_id
        type: integer
      - description: 用户名
        in: query
        name: username
        type: string
      - description: 操作类型
        in: query
        name: action
        type: string
      - description: 模块
        in: query
        name: module
        type: string
      - description: 状态码
        in: query
        name: status
        type: integer
      - description: 开始时间戳
        in: query
        name: start_time
        type: integer
      - description: 结束时间戳
        in: query
        name: end_time
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.SystemLogsResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取系统日志列表
      tags:
      - 系统日志
  /admin/logs/system/{id}:
    get:
      description: 根据ID获取单条系统日志的详细信息
      parameters:
      - description: 日志ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取系统日志详情
      tags:
      - 系统日志
  /admin/logs/system/stats:
    get:
      description: 统计指定时间范围内的系统日志数量
      parameters:
      - description: 开始时间戳
        in: query
        name: start_time
        type: integer
      - description: 结束时间戳
        in: query
        name: end_time
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取系统日志统计
      tags:
      - 系统日志
...
  /captcha:
    get:
      description: 获取验证码图片，返回验证码ID和Base64编码的图片
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.CaptchaResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取图形验证码
      tags:
      - 验证码
  /captcha/verify:
    post:
      consumes:
      - application/json
      description: 验证用户输入的验证码是否正确
      parameters:
      - description: 验证码信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.VerifyCaptchaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 验证验证码
      tags:
      - 验证码
...
securityDefinitions:
  BearerAuth:
    in: header
    name: token
    type: apiKey
swagger: "2.0"